Add static checker
For supporting the static checker, all macros part of the Domain
Specific Language (DSL) of libside must be defined with a level of
indirection.

`<side/static-check.h>', is included by `<side/api.h>' when
`SIDE_STATIC_CHECK_ENABLE' is defined.  The whole DSL is undefined and
redefined with a dispatching technique.  See the comments in the file
for details on how this works.

The static checker works by defining dummy C type and letting the
compiler found incompatible types at event definitions and call sites.

Change-Id: I7535a4829a6381039e66850ad825b9918626755d
Signed-off-by: Olivier Dion <odion@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>