Refactor description visitor
To ease refactoring in the future, all context arguments are wrapped
in a `visitor_context' structure.

This refactoring also introduce the `resolve_pointer_func' callback
while visiting.  This can be used while traversing events for pointers
that does not point to virtual memory.  For example, while traversing a
ELF file.

When no resolve callback is provided, pointers are not resolved.

Change-Id: I5d9766759dafce6421f2a9b7246f45eb2cd14ef8
Signed-off-by: Olivier Dion <odion@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>