C/C++ API: Set empty attribute list to NULL
If a list of attributes is empty, the compiler will assign an address
even though the object has a zero size. This address often aliases with
other objects and can not be interpreted as an attribute.

To help prevent user confusion when printing side attributes from a
debugger, set the attribute pointer to NULL when no attributes are
present.

Change-Id: I955b94b0ede00fc703b7145c1b7a39704ed43bae
Signed-off-by: Olivier Dion <odion@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>