Introduce side_array_t()
In multiple definitions, arrays were represented as a pointer to the
first element and a length.  For example `attr' and `nr_attr' in type
descriptions.  These were not contiguous.

However, this caused issues in the past with C++ that does not support
taking the address of a temporary array, even though Clang and GCC have
extensions that allow it.  These issues only involve none-static
compound literals, such as with dynamic attributes, thus the past
introduction of `side_dynamic_attr_list()'.  But because the head of the
array and its length were not contiguous, it was necessary to do macros
dispatching for selecting the correct expansion required for C++.

Furthermore, with GCC 13, a regression happened where "prvalue array
decay".  See the relevant bug report:
<https://gcc.gnu.org/bugzilla/show_bug.cgi?id=117817>.

This bug makes compound literals on static variables fail to compile for
GCC 13.  More specifically, this seems to only trigger an error when
passing the compound literal to the `sizeof' operator in
`SIDE_ARRAY_SIZE()'.  To overcome this issue, compound literals are
allocated with the `new' operator and their length is determined using a
private template helper called `libside::initializer_list_size<T>'.
However, this helper requires the type of the compound literal.

Therefore, the introduction of the `SIDE_LITERAL_ARRAY()' which expands
to a compound literal of a structure of the form:

	`struct { side_ptr_t(type) elements; uint32_t length; }'

To ensure that such initialization is possible, refactoring across all
structure with arrays is necessary.  Thus, the introduction of
`side_array_t' and the array accessor functions:

	- side_array_elements()
        - side_array_at()
        - side_array_length()
        - side_for_each_element_in_array()
        - side_array_index_of_element()

This solve the issue of compound literals in C++ and also make the code
more standardized across the project.

Note that since other toolchains could have similar problems, users can
define `SIDE_ALLOCATE_COMPOUND_LITERAL_ON_HEAP' to force using the
`libside::initializer_list_size<T>' helper.

Change-Id: Ia343a1d02818fd88556c6d3d65e08606eee3679e
Signed-off-by: Olivier Dion <odion@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>